+++ CUDADataFormats/SiPixelDigi/src/SiPixelDigisCUDA.cc
@@ -1,1 +1,3 @@
 #include "CUDADataFormats/SiPixelDigi/interface/SiPixelDigisCUDA.h"
+#include "CUDADataFormats/SiPixelDigi/interface/SiPixelDigisCUDAImpl.h"
+
