+++ HeterogeneousCore/CUDAUtilities/interface/cudaMemoryPoolImpl.h
@@ -4,6 +4,7 @@
 
 #include "HeterogeneousCore/CUDAUtilities/interface/cudaCheck.h"
 
+#include <cuda.h>
 #include <cuda_runtime.h>
 #include <cuda_runtime_api.h>
